fix: work with both the beta and app store versions
The app store version currently does not have the `ListWidget#addStack` method. This changes the widget to check whether it's availble, and use it if so. If not, it falls back to omitting the trend symbol.

I made this change because the Scriptable beta is full, and is likely to remain so. This condition can of course be removed once the app store version supports this API.