fix: don't crash if given a weird threshold
If threshold is a very low value, like -100, then the `find` call will return `undefined` and the destructuring assignment will fail, rather than using the default values as expected. To ensure this works, default to an empty object so all the defaults will be used, and fill in a default for `threshold`.